MB-20365 Reduce size of thread ids to 32bits
By reducing the thread ids to 4 bytes (from 8 bytes)
we reduce the size of a TraceEvent from 48 bytes to
40 bytes (Due to alignment).

Reviewed-on: http://review.couchbase.org/67207
Tested-by: buildbot <build@couchbase.com>

+++ include/phosphor/platform/thread.h
@@ -56,7 +56,7 @@ namespace phosphor {
          * @return thread id for the calling thread
          */
         PHOSPHOR_API
-        uint64_t getCurrentThreadID();
+        uint32_t getCurrentThreadID();
 
         /**
          * Get the cached system thread id for the calling thread
@@ -71,8 +71,8 @@ namespace phosphor {
          *
          * @return thread id for the calling thread
          */
-        inline uint64_t getCurrentThreadIDCached() {
-            static THREAD_LOCAL uint64_t thread_id;
+        inline uint32_t getCurrentThreadIDCached() {
+            static THREAD_LOCAL uint32_t thread_id;
 
             if (unlikely(!thread_id)) {
                 thread_id = getCurrentThreadID();
